requirements.txt
Flask==2.2.2
requests==2.28.1
numpy==1.23.4
Pillow==9.3.0
firebase-admin==5.3.0
google-cloud-storage==2.5.0
tensorflow==2.11.0
werkzeug==2.2.2
gunicorn